Data refactoring
