+++ .env.template
@@ -1,1 +1,3 @@
-export MY_DATASET_PATH="/home/jonny/datasets/blues"
+export YOUR_TRAIN_DATASET_PATH="/your/project/root/data/blues/train"
+export YOUR_VAL_DATASET_PATH="/your/project/root/data/blues/val"
+export YOUR_TEST_DATASET_PATH="/your/project/root/data/blues/test"
