+++ src/pl_data/datamodule.py
@@ -1,0 +1,82 @@
+from abc import abstractmethod
+from typing import Optional, Sequence
+
+import hydra
+import pytorch_lightning as pl
+from omegaconf import DictConfig
+from torch.utils.data import DataLoader, Dataset
+
+
+class MyDataModule(pl.LightningDataModule):
+    def __init__(
+            self,
+            datasets: DictConfig,
+            num_workers: DictConfig,
+            batch_size: DictConfig,
+            cfg: DictConfig,
+    ):
+        super().__init__()
+        self.cfg = cfg
+
+        self.datasets = datasets
+        self.num_workers = num_workers
+        self.batch_size = batch_size
+
+        self.train_dataset: Optional[Dataset] = None
+        self.val_datasets: Optional[Sequence[Dataset]] = None
+        self.test_datasets: Optional[Sequence[Dataset]] = None
+
+    def prepare_data(self) -> None:
+        # download only
+        pass
+
+    def setup(self, stage: Optional[str] = None):
+        # Here you should instantiate your datasets, you may also split the train into train and validation if needed.
+        if stage is None or stage == "fit":
+            self.train_dataset = hydra.utils.instantiate(self.datasets.train, cfg=self.cfg)
+            self.val_datasets = [
+                hydra.utils.instantiate(dataset_cfg, cfg=self.cfg) for dataset_cfg in self.datasets.val
+            ]
+
+        if stage is None or stage == "test":
+            self.test_datasets = [
+                hydra.utils.instantiate(dataset_cfg, cfg=self.cfg) for dataset_cfg in self.datasets.test
+            ]
+
+    def train_dataloader(self) -> DataLoader:
+        return DataLoader(
+            self.train_dataset,
+            shuffle=True,
+            batch_size=self.batch_size.train,
+            num_workers=self.num_workers.train,
+        )
+
+    def val_dataloader(self) -> Sequence[DataLoader]:
+        return [
+            DataLoader(
+                dataset,
+                shuffle=False,
+                batch_size=self.batch_size.val,
+                num_workers=self.num_workers.val,
+            )
+            for dataset in self.test_datasets
+        ]
+
+    def test_dataloader(self) -> Sequence[DataLoader]:
+        return [
+            DataLoader(
+                dataset,
+                shuffle=False,
+                batch_size=self.batch_size.test,
+                num_workers=self.num_workers.test,
+            )
+            for dataset in self.test_datasets
+        ]
+
+    def __repr__(self) -> str:
+        return (
+            f"{self.__class__.__name__}("
+            f"{self.datasets=}, "
+            f"{self.num_workers=}, "
+            f"{self.batch_size=})"
+        )
